fix startsAt and endsAt of ChimpagneEvent + refactor FindEventScreen
